backend/requirements.txt
 fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
numpy==2.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
opencv-python==4.10.0.84
matplotlib==3.9.2
scipy==1.14.1
scikit-image==0.24.0
requests==2.32.3
